medisense-ai-backend/requirements.txt
Flask
pymongo
Flask-Login
Flask-CORS
python-dotenv
openai